requirements.txt
selenium==4.15.0
python-telegram-bot==20.3
Pillow==10.0.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
apscheduler==3.10.4
gunicorn==21.2.0
